+++ src/main/java/io/discloader/discloader/network/json/RoleJSON.java
@@ -3,9 +3,9 @@ package io.discloader.discloader.network.json;
 public class RoleJSON {
 	public String id;
 	public String name;
-	public double permissions;
-	public double color;
-	public double position;
+	public long permissions;
+	public int color;
+	public int position;
 	public boolean hoist;
 	public boolean managed;
 	public boolean mentionable;
